Extracted step related operations from Flow to a new Step module
